Completed 'Studio: OMG more ORM!'
GET: /EventCategory/Create
POST: /EventCategory/Create

1. Controllers/EventCategoryController.cs
  - Handlers/action methods for adding event category
2. Models/EventCategory.cs, ViewModels/AddEventCatgoryViewModel.cs
  - View model for event category
3. Views/EventCategory/*
  - Form to get user information for event category
4. Views/Events/*, Views/Shared/*
  - Add links to different views
5. Controllers/EventsController.cs
  - a little extra error checking
6. Programs.cs, appsettings.json
  - use .json configuration for MySQL version numbers

+++ Controllers/EventsController.cs
@@ -62,10 +62,12 @@ public class EventsController : Controller
         foreach (int eventId in eventIds)
         {
             Event? theEvent = context.Events.Find(eventId);
-            context.Events.Remove(theEvent);
+            if (theEvent != null)
+            {
+                context.Events.Remove(theEvent);
+            }
         }
         context.SaveChanges();
-
         return Redirect("/Events");
     }
     // GET: /events/edit/{id}
